requirements.txt
numpy==1.22.3
Pillow==10.2.0
streamlit==1.26.0
tensorflow==2.9.1
